+++ src/view_modes/markdown.rs
@@ -14,7 +14,8 @@ fn pre_render_html(md: &str) -> String {
     static ITALIC: LazyLock<Regex> =
         LazyLock::new(|| Regex::new(r"(?is)<(?:i|em)>(.*?)</(?:i|em)>").unwrap());
     static UNDERLINE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(?is)<u>(.*?)</u>").unwrap());
-    static CODE: LazyLock<Regex> = LazyLock::new(|| Regex::new(r"(?is)<code>(.*?)</code>").unwrap());
+    static CODE: LazyLock<Regex> =
+        LazyLock::new(|| Regex::new(r"(?is)<code>(.*?)</code>").unwrap());
     static LINK: LazyLock<Regex> = LazyLock::new(|| {
         Regex::new(r#"(?is)<a\s+[^>]*href\s*=\s*["']([^"']*)["'][^>]*>(.*?)</a>"#).unwrap()
     });
@@ -130,11 +131,13 @@ pub fn render_markdown_view(ui: &mut egui::Ui, tab: &mut TabState) {
             });
         });
 
-        let clicked_fragment = tab
-            .commonmark_cache
-            .link_hooks()
-            .iter()
-            .find_map(|(k, v)| if *v { Some(k.clone()) } else { None });
+        let clicked_fragment = tab.commonmark_cache.link_hooks().iter().find_map(|(k, v)| {
+            if *v {
+                Some(k.clone())
+            } else {
+                None
+            }
+        });
         if let Some(frag) = clicked_fragment {
             tab.commonmark_cache
                 .link_hooks_mut()